DEV :: 채팅 초대, 추방, 나가기, 생성까지 끝

+++ src/main/java/nuri/nuri_server/global/properties/ChatProperties.java
@@ -8,5 +8,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @RequiredArgsConstructor
 @ConfigurationProperties(prefix = "chat")
 public class ChatProperties {
-    private final Long broadcastThreshold;
+    private final Integer broadcastThreshold;
 }
